Do not track Gemfile.lock
A Gemfile.lock just gets in the way in a gem. We typically want to test against the latest versions
of dependencies (as constrained by our .gemspec), not be locked to any specific version. So it
should be safe and easy to just `rm Gemfile.lock` and use the latest versions at any time.

See also 3fd1c68c334def332c888bc9d1850a0dc7ebf452