[#7] Write specs (#84)
* [#7] Write specs

* make tylers pr mergeable

* Tyler's work (#83)

* Don't strictly require Rails >= 5.2

Allow gems to be used with Rails 5.1 as well

* Configure for testing

- Add combustion and rspec-rails as development dependencies
- Add files generated by `combustion` generator
- Add Combustion to spec_helper.rb

Update .travis.yml to run tests in all sub-projects too
- Extracted those steps as bin/test_all so we can easily run it in development too

* Do not track Gemfile.lock

A Gemfile.lock just gets in the way in a gem. We typically want to test against the latest versions
of dependencies (as constrained by our .gemspec), not be locked to any specific version. So it
should be safe and easy to just `rm Gemfile.lock` and use the latest versions at any time.

See also 3fd1c68c334def332c888bc9d1850a0dc7ebf452

* Fix the require for engine

Fixes an error encountered when requiring from tests:

  Failure/Error: require 'notification-handler'

  NameError:
    uninitialized constant Rails::Engine

(require 'rails/engine' will in turn call require 'rails/railtie')

Don't need separate file for an Engine and a Railtie because an Engine *is* a Railtie

* [#7] Add some automated tests!

Includes both:

1. Tests for each gem in isolation (with only the absolute minimum dependencies required for it to
work):

- notification-handler/spec/
- notification-pusher/spec/
- notification-settings/spec/

2. Tests that test all gems loaded together (like if you added the 'notifications-rails' gem to your
project):

- spec/

* [#57] Fix: Using user.notify inside a transaction can cause same pusher to get called twice

Resolves #57

* Rename initialize_pusher to push_using_pusher_from_attribute to make it clear that it doesn't just
  initialize something — it actually does a push if it can (if you have told it which pusher you
  want)!

* Replace initiate_push method with push and push!

* Change find_by_name to return single pusher instead of an array (since we were always calling
  .first on the resulting array)

* Add find_by_name!

* [#44] Pass a callable to `define_group` so evaluation of relation is deferred

If you try to use a model from within NotificationHandler.configure:

  NotificationHandler.configure do |config|
    config.define_group :subscribers, User.where(subscriber: true)
  end

it would trigger the model to be loaded, but if it referenced a method from NotificationHandler:

  class User < ApplicationRecord
    notification_target
  end

then it created a circular dependency and resulted in an error such as:

  NameError:
    undefined local variable or method `notification_target' for User (call 'User.connection' to establish a connection):Class

Resolves #44

* Fix error when calling Subscriber#subscribe:

  NoMethodError:
    undefined method `create!' for nil:NilClass

    18: def create_notification_setting
 => 19:   notification_setting.create!
    20: end

Change to use before_create :build_notification_setting, and automatically build
`notification_setting` if object doesn't have one yet like we did in #40.

* * Fix: settings_allow_push?(:Name) ignored setting.settings[:Name] = false

local_pusher_settings?(pusher) was checking:

  local_settings_allow_push?(pusher).present?

But false.present? is false so when local setting was {:Name => false},
local_pusher_settings?(:Name) was returning false even though they key was present and the local
setting should have been overriding the global setting.

Fixed by changing check to:

  !local_pusher_setting(pusher).nil?

Similar change in local_pusher_category_settings?

* Raise more useful errors in case the requested pusher was nil or was not found

* Remove duplication with the dig call by extracting category_setting

* Test Ruby 2.6

* [#50] Handle/validate missing Notification#target

* Improve Changelog entry about pushers needing a `call` method

* add Gemfile.lock

* some changes

* more changes

* remove Gemfile.lock

* update spec

* run all rspec tests

* update specs

* fix automated tests

* add specs to pusher gems